+++ src/main/java/com/buuz135/functionalstorage/FunctionalStorage.java
@@ -4,10 +4,7 @@ import com.buuz135.functionalstorage.block.*;
 import com.buuz135.functionalstorage.block.tile.*;
 import com.buuz135.functionalstorage.client.*;
 import com.buuz135.functionalstorage.client.loader.FramedModel;
-import com.buuz135.functionalstorage.data.FunctionalStorageBlockTagsProvider;
-import com.buuz135.functionalstorage.data.FunctionalStorageBlockstateProvider;
-import com.buuz135.functionalstorage.data.FunctionalStorageItemTagsProvider;
-import com.buuz135.functionalstorage.data.FunctionalStorageLangProvider;
+import com.buuz135.functionalstorage.data.*;
 import com.buuz135.functionalstorage.inventory.BigInventoryHandler;
 import com.buuz135.functionalstorage.inventory.item.CompactingStackItemHandler;
 import com.buuz135.functionalstorage.inventory.item.DrawerStackItemHandler;
@@ -16,32 +13,35 @@ import com.buuz135.functionalstorage.item.LinkingToolItem;
 import com.buuz135.functionalstorage.item.StorageUpgradeItem;
 import com.buuz135.functionalstorage.item.UpgradeItem;
 import com.buuz135.functionalstorage.network.EnderDrawerSyncMessage;
+import com.buuz135.functionalstorage.recipe.CustomCompactingRecipe;
 import com.buuz135.functionalstorage.recipe.DrawerlessWoodIngredient;
 import com.buuz135.functionalstorage.recipe.FramedDrawerRecipe;
-import com.buuz135.functionalstorage.util.*;
-import com.hrznstudio.titanium.block.BasicBlock;
+import com.buuz135.functionalstorage.util.DrawerWoodType;
+import com.buuz135.functionalstorage.util.IWoodType;
+import com.buuz135.functionalstorage.util.NumberUtils;
+import com.buuz135.functionalstorage.util.TooltipUtil;
 import com.hrznstudio.titanium.datagenerator.loot.TitaniumLootTableProvider;
 import com.hrznstudio.titanium.datagenerator.model.BlockItemModelGeneratorProvider;
 import com.hrznstudio.titanium.event.handler.EventManager;
 import com.hrznstudio.titanium.module.ModuleController;
 import com.hrznstudio.titanium.nbthandler.NBTManager;
 import com.hrznstudio.titanium.network.NetworkHandler;
-import com.hrznstudio.titanium.recipe.generator.TitaniumRecipeProvider;
-import com.hrznstudio.titanium.recipe.generator.TitaniumShapedRecipeBuilder;
+import com.hrznstudio.titanium.recipe.generator.IJSONGenerator;
+import com.hrznstudio.titanium.recipe.generator.IJsonFile;
+import com.hrznstudio.titanium.recipe.generator.TitaniumSerializableProvider;
+import com.hrznstudio.titanium.recipe.serializer.GenericSerializer;
 import com.hrznstudio.titanium.tab.AdvancedTitaniumTab;
 import net.minecraft.client.renderer.ItemBlockRenderTypes;
 import net.minecraft.client.renderer.RenderType;
-import net.minecraft.data.recipes.FinishedRecipe;
-import net.minecraft.data.recipes.UpgradeRecipeBuilder;
 import net.minecraft.data.tags.BlockTagsProvider;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.resources.ResourceLocation;
-import net.minecraft.tags.ItemTags;
 import net.minecraft.world.inventory.InventoryMenu;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.Items;
-import net.minecraft.world.item.crafting.Ingredient;
+import net.minecraft.world.item.crafting.RecipeSerializer;
+import net.minecraft.world.item.crafting.RecipeType;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.level.block.entity.BlockEntityType;
@@ -52,7 +52,6 @@ import net.minecraftforge.client.event.*;
 import net.minecraftforge.client.model.generators.BlockModelProvider;
 import net.minecraftforge.client.model.generators.ItemModelProvider;
 import net.minecraftforge.common.ForgeMod;
-import net.minecraftforge.common.Tags;
 import net.minecraftforge.common.capabilities.ForgeCapabilities;
 import net.minecraftforge.common.crafting.CraftingHelper;
 import net.minecraftforge.common.util.NonNullLazy;
@@ -71,7 +70,6 @@ import org.apache.logging.log4j.Logger;
 import java.awt.*;
 import java.util.List;
 import java.util.*;
-import java.util.function.Consumer;
 import java.util.function.Function;
 import java.util.stream.Collectors;
 
@@ -116,6 +114,9 @@ public class FunctionalStorage extends ModuleController {
     public static RegistryObject<Item> CONFIGURATION_TOOL;
     public static RegistryObject<Item> REDSTONE_UPGRADE;
     public static RegistryObject<Item> CREATIVE_UPGRADE;
+
+    public static RegistryObject<RecipeSerializer<?>> CUSTOM_COMPACTING_RECIPE_SERIALIZER;
+    public static RegistryObject<RecipeType<?>> CUSTOM_COMPACTING_RECIPE_TYPE;
 
     public static AdvancedTitaniumTab TAB = new AdvancedTitaniumTab("functionalstorage", true);
 
@@ -229,6 +230,10 @@ public class FunctionalStorage extends ModuleController {
         FRAMED_CONTROLLER_EXTENSION = getRegistries().registerBlockWithTile("framed_controller_extension", FramedControllerExtensionBlock::new);
 
         getRegistries().registerGeneric(ForgeRegistries.RECIPE_SERIALIZERS.getRegistryKey(), "framed_recipe", () -> FramedDrawerRecipe.SERIALIZER);
+
+        CUSTOM_COMPACTING_RECIPE_TYPE = getRegistries().registerGeneric(ForgeRegistries.RECIPE_TYPES.getRegistryKey(), "custom_compacting", () -> RecipeType.simple(new ResourceLocation(MOD_ID, "custom_compacting")));
+
+        CUSTOM_COMPACTING_RECIPE_SERIALIZER = getRegistries().registerGeneric(ForgeRegistries.RECIPE_SERIALIZERS.getRegistryKey(), "custom_compacting", () -> new GenericSerializer<CustomCompactingRecipe>((RecipeType<CustomCompactingRecipe>) CUSTOM_COMPACTING_RECIPE_TYPE.get(), CustomCompactingRecipe.class));
     }
 
     public enum DrawerType {
@@ -442,99 +447,19 @@ public class FunctionalStorage extends ModuleController {
                 }
             });
         }
-        event.getGenerator().addProvider(true, new TitaniumRecipeProvider(event.getGenerator()) {
+        event.getGenerator().addProvider(true, new FunctionalStorageRecipesProvider(event.getGenerator(), blocksToProcess));
+        event.getGenerator().addProvider(true, new TitaniumSerializableProvider(event.getGenerator(), MOD_ID) {
             @Override
-            public void register(Consumer<FinishedRecipe> consumer) {
-                blocksToProcess.get().stream().map(block -> (BasicBlock) block).forEach(basicBlock -> basicBlock.registerRecipe(consumer));
-                TitaniumShapedRecipeBuilder.shapedRecipe(STORAGE_UPGRADES.get(StorageUpgradeItem.StorageTier.IRON).get())
-                        .pattern("III").pattern("IDI").pattern("III")
-                        .define('I', Tags.Items.INGOTS_IRON)
-                        .define('D', StorageTags.DRAWER)
-                        .save(consumer);
-                TitaniumShapedRecipeBuilder.shapedRecipe(VOID_UPGRADE.get())
-                        .pattern("III").pattern("IDI").pattern("III")
-                        .define('I', Tags.Items.OBSIDIAN)
-                        .define('D', StorageTags.DRAWER)
-                        .save(consumer);
-                TitaniumShapedRecipeBuilder.shapedRecipe(CONFIGURATION_TOOL.get())
-                        .pattern("PPG").pattern("PDG").pattern("PEP")
-                        .define('P', Items.PAPER)
-                        .define('G', Tags.Items.INGOTS_GOLD)
-                        .define('D', StorageTags.DRAWER)
-                        .define('E', Items.EMERALD)
-                        .save(consumer);
-                TitaniumShapedRecipeBuilder.shapedRecipe(LINKING_TOOL.get())
-                        .pattern("PPG").pattern("PDG").pattern("PEP")
-                        .define('P', Items.PAPER)
-                        .define('G', Tags.Items.INGOTS_GOLD)
-                        .define('D', StorageTags.DRAWER)
-                        .define('E', Items.DIAMOND)
-                        .save(consumer);
-                TitaniumShapedRecipeBuilder.shapedRecipe(STORAGE_UPGRADES.get(StorageUpgradeItem.StorageTier.COPPER).get())
-                        .pattern("IBI").pattern("CDC").pattern("IBI")
-                        .define('I', Items.COPPER_INGOT)
-                        .define('B', Items.COPPER_BLOCK)
-                        .define('C', Tags.Items.CHESTS_WOODEN)
-                        .define('D', StorageTags.DRAWER)
-                        .save(consumer);
-                TitaniumShapedRecipeBuilder.shapedRecipe(STORAGE_UPGRADES.get(StorageUpgradeItem.StorageTier.GOLD).get())
-                        .pattern("IBI").pattern("CDC").pattern("BIB")
-                        .define('I', Tags.Items.INGOTS_GOLD)
-                        .define('B', Tags.Items.STORAGE_BLOCKS_GOLD)
-                        .define('C', Tags.Items.CHESTS_WOODEN)
-                        .define('D', STORAGE_UPGRADES.get(StorageUpgradeItem.StorageTier.COPPER).get())
-                        .save(consumer);
-                TitaniumShapedRecipeBuilder.shapedRecipe(STORAGE_UPGRADES.get(StorageUpgradeItem.StorageTier.DIAMOND).get())
-                        .pattern("IBI").pattern("CDC").pattern("IBI")
-                        .define('I', Tags.Items.GEMS_DIAMOND)
-                        .define('B', Tags.Items.STORAGE_BLOCKS_DIAMOND)
-                        .define('C', Tags.Items.CHESTS_WOODEN)
-                        .define('D', STORAGE_UPGRADES.get(StorageUpgradeItem.StorageTier.GOLD).get())
-                        .save(consumer);
-                TitaniumShapedRecipeBuilder.shapedRecipe(REDSTONE_UPGRADE.get())
-                        .pattern("IBI").pattern("CDC").pattern("IBI")
-                        .define('I', Items.REDSTONE)
-                        .define('B', Items.REDSTONE_BLOCK)
-                        .define('C', Items.COMPARATOR)
-                        .define('D', StorageTags.DRAWER)
-                        .save(consumer);
-                UpgradeRecipeBuilder.smithing(Ingredient.of(STORAGE_UPGRADES.get(StorageUpgradeItem.StorageTier.DIAMOND).get()), Ingredient.of(Items.NETHERITE_INGOT), STORAGE_UPGRADES.get(StorageUpgradeItem.StorageTier.NETHERITE).get())
-                        .unlocks("has_netherite_ingot", has(Items.NETHERITE_INGOT))
-                        .save(consumer, ForgeRegistries.ITEMS.getKey(STORAGE_UPGRADES.get(StorageUpgradeItem.StorageTier.NETHERITE).get()));
-                TitaniumShapedRecipeBuilder.shapedRecipe(ARMORY_CABINET.getLeft().get())
-                        .pattern("ICI").pattern("CDC").pattern("IBI")
-                        .define('I', Tags.Items.STONE)
-                        .define('B', Tags.Items.INGOTS_NETHERITE)
-                        .define('C', StorageTags.DRAWER)
-                        .define('D', Items.COMPARATOR)
-                        .save(consumer);
-                TitaniumShapedRecipeBuilder.shapedRecipe(PULLING_UPGRADE.get())
-                        .pattern("ICI").pattern("IDI").pattern("IBI")
-                        .define('I', Tags.Items.STONE)
-                        .define('B', Tags.Items.DUSTS_REDSTONE)
-                        .define('C', Items.HOPPER)
-                        .define('D', StorageTags.DRAWER)
-                        .save(consumer);
-                TitaniumShapedRecipeBuilder.shapedRecipe(PUSHING_UPGRADE.get())
-                        .pattern("IBI").pattern("IDI").pattern("IRI")
-                        .define('I', Tags.Items.STONE)
-                        .define('B', Tags.Items.DUSTS_REDSTONE)
-                        .define('R', Items.HOPPER)
-                        .define('D', StorageTags.DRAWER)
-                        .save(consumer);
-                TitaniumShapedRecipeBuilder.shapedRecipe(COLLECTOR_UPGRADE.get())
-                        .pattern("IBI").pattern("RDR").pattern("IBI")
-                        .define('I', Tags.Items.STONE)
-                        .define('B', Items.HOPPER)
-                        .define('R', Tags.Items.DUSTS_REDSTONE)
-                        .define('D', StorageTags.DRAWER)
-                        .save(consumer);
-                TitaniumShapedRecipeBuilder.shapedRecipe(ENDER_DRAWER.getLeft().get())
-                        .pattern("PPP").pattern("LCL").pattern("PPP")
-                        .define('P', ItemTags.PLANKS)
-                        .define('C', Tags.Items.CHESTS_ENDER)
-                        .define('L', StorageTags.DRAWER)
-                        .save(consumer);
+            public void add(Map<IJsonFile, IJSONGenerator> serializables) {
+                new CustomCompactingRecipe(new ResourceLocation("clay"), new ItemStack(Items.CLAY_BALL, 4), new ItemStack(Items.CLAY));
+                new CustomCompactingRecipe(new ResourceLocation("glowstone"), new ItemStack(Items.GLOWSTONE_DUST, 4), new ItemStack(Items.GLOWSTONE));
+                new CustomCompactingRecipe(new ResourceLocation("melon"), new ItemStack(Items.MELON_SLICE, 9), new ItemStack(Items.MELON));
+                new CustomCompactingRecipe(new ResourceLocation("quartz"), new ItemStack(Items.QUARTZ, 4), new ItemStack(Items.QUARTZ_BLOCK));
+                new CustomCompactingRecipe(new ResourceLocation("ice"), new ItemStack(Items.ICE, 9), new ItemStack(Items.BLUE_ICE));
+                new CustomCompactingRecipe(new ResourceLocation("blue_ice"), new ItemStack(Items.BLUE_ICE, 9), new ItemStack(Items.PACKED_ICE));
+                new CustomCompactingRecipe(new ResourceLocation("amethyst"), new ItemStack(Items.AMETHYST_SHARD, 9), new ItemStack(Items.AMETHYST_BLOCK));
+
+                CustomCompactingRecipe.RECIPES.forEach(customCompactingRecipe -> serializables.put(customCompactingRecipe, customCompactingRecipe));
             }
         });
     }